feat: create store and components
