+++ homeworks/edelmiro-anton/20-routing-redux/redux/todo-app/package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.0.4",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
