+++ homeworks/edelmiro-anton/20-routing-redux/redux/todo-app/src/App.css
@@ -5,50 +5,4 @@
   background: linear-gradient(116.67deg, #8c72b7 0.58%, rgba(192, 46, 195, 0) 101.02%);
 }
 
-.mainInput {
-  display: flex;
-  justify-content: center;
-  margin-top: 99px;
-}
 
-.input_todoTask {
-  width: 554px;
-  margin-top: 99px;
-  background: #ecf3ff;
-  border: none;
-  font-family: 'Hind Siliguri';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 32px;
-  line-height: 52px;
-  color: rgba(0, 0, 0, 0.6);
-}
-
-::placeholder {
-  width: 239px;
-  height: 52px;
-  padding-left: 100px;
-  text-align: center;
-  font-family: 'Hind Siliguri';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 32px;
-  line-height: 52px;
-  color: rgba(0, 0, 0, 0.6);
-}
-
-.button_add {
-  width: 122px;
-  height: 65px;
-  left: 958px;
-  margin-top: 99px;
-  background: #5275a0;
-  font-family: 'Hind Siliguri';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 32px;
-  line-height: 52px;
-  color: #ffffff;
-  border: none;
-  cursor: pointer;
-}
