+++ homeworks/edelmiro-anton/20-routing-redux/redux/todo-app/src/App.js
@@ -1,12 +1,10 @@
+import { AddTodo } from './components/AddTodo';
 import './App.css';
 
 function App() {
   return (
     <div className="app">
-      <div className="mainInput">
-        <input type="text" className="input_todoTask" placeholder="Create Todo-Task" />
-        <button className="button_add">Add</button>
-      </div>
+      <AddTodo />
     </div>
   );
 }
