+++ homeworks/edelmiro-anton/20-routing-redux/redux/todo-app/src/components/AddTodo.js
@@ -1,0 +1,12 @@
+import React, { useState } from 'react';
+
+import './AddTodo.css';
+
+export const AddTodo = () => {
+  return (
+    <div className="mainInput">
+      <input type="text" className="input_todoTask" placeholder="Create Todo-Task" />
+      <button className="button_add">Add</button>
+    </div>
+  );
+};
