+++ homeworks/edelmiro-anton/20-routing-redux/redux/todo-app/src/components/TodoInput.css
@@ -1,0 +1,29 @@
+.todo_container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 23px;
+}
+
+.input_todo {
+  width: 554px;
+  margin-top: auto;
+  background: #ecf3ff;
+  border: none;
+  font-family: 'Hind Siliguri';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 32px;
+  line-height: 52px;
+  color: rgba(0, 0, 0, 0.6);
+}
+
+.delete_icon {
+  width: 59.38px;
+  height: 65px;
+}
+
+.edit_icon {
+  width: 63.23px;
+  height: 65px;
+}
