--- homeworks/edelmiro-anton/20-routing-redux/redux/todo-app/src/store.js
@@ -1,6 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit'
-
-
-export const store = configureStore({
-    reducer: {},
-  })
