+++ homeworks/edelmiro-anton/20-routing-redux/redux/todo-app/src/store/index.js
@@ -1,0 +1,5 @@
+import { configureStore } from '@reduxjs/toolkit';
+
+export const store = configureStore({
+  reducer: {},
+});
